fix(data): handle fractional sgv values from Nightscout entries
Juggluco (and other CGM sources, e.g. Libre 2/2+) send sgv as a float
without rounding. EntryDto declared sgv as Int, so Moshi threw a JSON
parse error on any fractional value ("Expected an int but was
136.86...").

Change sgv to Double throughout the entry pipeline (EntryDto,
GlucoseEntry, GlucoseAlertManager, GlucoseGraphRenderer) and round
only at display/notification time for mg/dL. mmol/L display already
used floating point math and is unaffected.

Fixes #13

+++ app/src/main/java/de/autosugar/car/GlucoseScreen.kt
@@ -138,13 +138,13 @@ class GlucoseScreen(
         if (projected15 > thresholds.bgHigh && sgv <= thresholds.bgHigh &&
             now - lastPredictedHighAlertMs > alertCooldownMs
         ) {
-            alertManager.sendPredictedHighAlert(projected15.toInt(), profile.unit)
+            alertManager.sendPredictedHighAlert(projected15, profile.unit)
             lastPredictedHighAlertMs = now
         }
         if (projected15 < thresholds.bgLow && sgv >= thresholds.bgLow &&
             now - lastPredictedLowAlertMs > alertCooldownMs
         ) {
-            alertManager.sendPredictedLowAlert(projected15.toInt(), profile.unit)
+            alertManager.sendPredictedLowAlert(projected15, profile.unit)
             lastPredictedLowAlertMs = now
         }
     }
